refactor: delegate waiting entry logic to WaitingService and clean up AdminService
