+++ backend/src/main/java/com/example/demo/application/service/AdminService.java
@@ -13,14 +13,12 @@ import com.example.demo.domain.model.waiting.WaitingQuery;
 import com.example.demo.domain.model.waiting.WaitingStatus;
 import com.example.demo.domain.port.PopupPort;
 import com.example.demo.domain.port.WaitingPort;
-import com.example.demo.domain.port.WaitingStatisticsPort;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @Slf4j
@@ -31,11 +29,10 @@ public class AdminService {
     private final AppProperties appProperties;
     private final PopupPort popupPort;
     private final WaitingPort waitingPort;
-    private final WaitingStatisticsPort waitingStatisticsPort;
     private final AdminSessionService adminSessionService;
     private final PopupService popupService;
     private final ImageService imageService;
-    private final WaitingNotificationService waitingNotificationService;
+    private final WaitingService waitingService;
 
     /**
      * 관리자 비밀번호 검증 및 세션 토큰 발급
@@ -158,93 +155,15 @@ public class AdminService {
     }
 
     /**
-     * 대기 입장 처리 (맨 앞 순번만 가능)
+     * 대기 입장 처리 (관리자용)
+     * 실제 입장 처리 로직은 WaitingService에 위임
      */
     @Transactional
     public void enterWaiting(Long waitingId) {
-        Waiting waiting = waitingPort.findByQuery(WaitingQuery.forWaitingId(waitingId))
-                .stream()
-                .findFirst()
-                .orElseThrow(() -> new BusinessException(ErrorType.WAITING_NOT_FOUND, String.valueOf(waitingId)));
-
-        // 대기 번호가 0번인지 확인
-        if (waiting.waitingNumber() != 0) {
-            throw new BusinessException(ErrorType.WAITING_NOT_READY);
-        }
-
-        // 상태가 WAITING인지 확인
-        if (waiting.status() != WaitingStatus.WAITING) {
-            throw new BusinessException(ErrorType.INVALID_WAITING_STATUS, waiting.status().name());
-        }
-
-        // 입장 처리
-        Waiting enteredWaiting = waiting.enter();
-        waitingPort.save(enteredWaiting);
-
-        // 나머지 대기자들의 번호 감소
-        decrementWaitingNumbers(waiting.popup().getId());
-
-        log.info("대기 입장 처리 완료: waitingId={}, member={}", waitingId, waiting.member().name());
-    }
-
-    /**
-     * 대기 번호 감소 (입장 처리 후)
-     * N+1 문제 방지를 위해 배치 저장 사용
-     */
-    private void decrementWaitingNumbers(Long popupId) {
-        WaitingQuery query = WaitingQuery.forPopup(popupId, WaitingStatus.WAITING);
-        List<Waiting> waitings = waitingPort.findByQuery(query);
+        // WaitingService에 위임 (도메인 로직 재사용)
+        waitingService.enterWaiting(waitingId);
         
-        // 팝업 통계 조회
-        var statistics = waitingStatisticsPort.findCompletedStatisticsByPopupId(popupId);
-        Double avgTimePerPerson = statistics.calculateAverageTimePerPerson();
-        log.info("[입장 처리] 예상 대기 시간 업데이트 시작 - popupId: {}, 평균 대기시간: {}분/팀, 대기자 수: {}", 
-                popupId, avgTimePerPerson, waitings.size());
-
-        Waiting newFirstWaiting = null;
-        Waiting new3rdWaiting = null;
-        List<Waiting> decrementedWaitings = new ArrayList<>();
-
-        // 대기 중인 모든 대기자의 번호를 1씩 감소하고 리스트에 모음
-        for (Waiting waiting : waitings) {
-            if (waiting.waitingNumber() > 0) {
-                Waiting decremented = waiting.minusWaitingNumber(statistics);
-                decrementedWaitings.add(decremented);
-                
-                log.debug("[입장 처리] 예상 대기 시간 업데이트 - waitingId: {}, 대기번호: {}번->{}번, 예상시간: {}분", 
-                        waiting.id(), waiting.waitingNumber(), decremented.waitingNumber(), 
-                        decremented.expectedWaitingTimeMinutes());
-                
-                // 새로 0번이 된 사람 (기존 1번)
-                if (decremented.waitingNumber() == 0) {
-                    newFirstWaiting = decremented;
-                } 
-                // 새로 3번이 된 사람 (기존 4번)
-                else if (decremented.waitingNumber() == 3) {
-                    new3rdWaiting = decremented;
-                }
-            }
-        }
-        
-        // 배치로 한 번에 저장 (N+1 문제 해결)
-        if (!decrementedWaitings.isEmpty()) {
-            waitingPort.saveAll(decrementedWaitings);
-            log.info("[입장 처리] 예상 대기 시간 업데이트 완료 - {} 건 일괄 저장", decrementedWaitings.size());
-        }
-        
-        // 새로 0번이 된 사람에게 입장 알림 발송 (SSE + 이메일)
-        if (newFirstWaiting != null) {
-            waitingNotificationService.sendEnterNowNotification(newFirstWaiting);
-            log.info("새로운 0번 대기자에게 입장 알림 발송: waitingId={}, memberId={}", 
-                    newFirstWaiting.id(), newFirstWaiting.member().id());
-        }
-        
-        // 새로 3번이 된 사람에게 3팀 전 알림 발송 (SSE)
-        if (new3rdWaiting != null) {
-            waitingNotificationService.sendEnter3TeamsBeforeNotification(new3rdWaiting);
-            log.info("3팀 전 알림 발송: waitingId={}, memberId={}", 
-                    new3rdWaiting.id(), new3rdWaiting.member().id());
-        }
+        log.info("[Admin] 대기 입장 처리 완료: waitingId={}", waitingId);
     }
 
     /**
