+++ backend/src/test/java/com/example/demo/application/service/WaitingServiceTest.java
@@ -683,9 +683,9 @@ class WaitingServiceTest {
         }
     }
 
-    @DisplayName("makeVisit 메서드 테스트")
+    @DisplayName("enterWaiting 메서드 테스트")
     @Nested
-    class MakeVisitTest {
+    class EnterWaitingTest {
         private final Popup validPopup = Popup.builder()
                 .id(1L)
                 .name("테스트 팝업")
@@ -735,7 +735,7 @@ class WaitingServiceTest {
 
         @Test
         @DisplayName("존재하지 않는 대기 ID로 입장 처리 시 WAITING_NOT_FOUND 예외")
-        void makeVisit_WaitingNotFound() {
+        void enterWaiting_WaitingNotFound() {
             // given
             Long waitingId = 999L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -745,7 +745,7 @@ class WaitingServiceTest {
 
             // when & then
             BusinessException exception = assertThrows(BusinessException.class,
-                    () -> waitingService.makeVisit(request));
+                    () -> waitingService.enterWaiting(request.waitingId()));
 
             assertEquals(ErrorType.WAITING_NOT_FOUND, exception.getErrorType());
             assertEquals(String.valueOf(waitingId), exception.getAdditionalInfo());
@@ -753,7 +753,7 @@ class WaitingServiceTest {
 
         @Test
         @DisplayName("이미 방문 완료된 대기에서 입장 처리 시 INVALID_WAITING_STATUS 예외")
-        void makeVisit_AlreadyVisited() {
+        void enterWaiting_AlreadyVisited() {
             // given
             Long waitingId = 1L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -769,14 +769,14 @@ class WaitingServiceTest {
 
             // when & then
             BusinessException exception = assertThrows(BusinessException.class,
-                    () -> waitingService.makeVisit(request));
+                    () -> waitingService.enterWaiting(request.waitingId()));
 
             assertEquals(ErrorType.INVALID_WAITING_STATUS, exception.getErrorType());
         }
 
         @Test
         @DisplayName("취소된 대기에서 입장 처리 시 INVALID_WAITING_STATUS 예외")
-        void makeVisit_CanceledWaiting() {
+        void enterWaiting_CanceledWaiting() {
             // given
             Long waitingId = 1L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -792,14 +792,14 @@ class WaitingServiceTest {
 
             // when & then
             BusinessException exception = assertThrows(BusinessException.class,
-                    () -> waitingService.makeVisit(request));
+                    () -> waitingService.enterWaiting(request.waitingId()));
 
             assertEquals(ErrorType.INVALID_WAITING_STATUS, exception.getErrorType());
         }
 
         @Test
         @DisplayName("중간 순번 입장 시 WAITING_NOT_READY 예외 발생")
-        void makeVisit_ReorderMiddleNumber() {
+        void enterWaiting_ReorderMiddleNumber() {
             // given
             Long waitingId = 3L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -838,14 +838,14 @@ class WaitingServiceTest {
 
             // when & then
             BusinessException exception = assertThrows(BusinessException.class,
-                    () -> waitingService.makeVisit(request));
+                    () -> waitingService.enterWaiting(request.waitingId()));
 
             assertEquals(ErrorType.WAITING_NOT_READY, exception.getErrorType());
         }
 
         @Test
         @DisplayName("첫 번째 순번 입장 시 모든 뒤 순번들이 앞당겨짐")
-        void makeVisit_ReorderFirstNumber() {
+        void enterWaiting_ReorderFirstNumber() {
             // given
             Long waitingId = 1L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -887,7 +887,7 @@ class WaitingServiceTest {
                     .thenReturn(List.of(waiting2, waiting3, waiting4));
 
             // when
-            waitingService.makeVisit(request);
+            waitingService.enterWaiting(request.waitingId());
 
             // then
             // 입장 처리된 대기 1번 save + 순번 변경된 대기들 saveAll 1번 = 총 2번 호출
@@ -898,7 +898,7 @@ class WaitingServiceTest {
 
         @Test
         @DisplayName("단일 대기 입장 시 순번 재배치 없음")
-        void makeVisit_SingleWaiting() {
+        void enterWaiting_SingleWaiting() {
             // given
             Long waitingId = 1L;
             WaitingMakeVisitRequest request = new WaitingMakeVisitRequest(waitingId);
@@ -915,13 +915,16 @@ class WaitingServiceTest {
             when(waitingPort.findByQuery(WaitingQuery.forPopup(validPopup.getId(), WaitingStatus.WAITING)))
                     .thenReturn(List.of()); // 다른 대기 없음
 
+            when(waitingStatisticsPort.findCompletedStatisticsByPopupId(validPopup.getId()))
+                    .thenReturn(new PopupWaitingStatistics(validPopup.getId(), List.of()));
+
             // when
-            waitingService.makeVisit(request);
+            waitingService.enterWaiting(request.waitingId());
 
             // then
-            // 입장 처리된 대기 1번 save + 빈 리스트로 saveAll 1번
+            // 입장 처리된 대기 1번 save (다른 대기가 없으므로 saveAll은 호출되지 않음)
             verify(waitingPort, times(1)).save(any(Waiting.class));
-            verify(waitingPort, times(1)).saveAll(List.of());
+            verify(waitingPort, never()).saveAll(anyList());
         }
     }
 }
